Functions In Progress

+++ src/sample/pair.java
@@ -2,13 +2,13 @@ package sample;
 
 public class pair
 {
-    private int X;
-    private int Y;
+    private double X;
+    private double Y;
 
-    public pair(int x, int y) { X = x; Y = y; }
+    public pair(double x, double y) { X = x; Y = y; }
 
-    public int getX() { return X; }
-    public void setX(int x) { X = x; }
-    public int getY() { return Y; }
-    public void setY(int y) { Y = y; }
+    public double getX() { return X; }
+    public void setX(double x) { X = x; }
+    public double getY() { return Y; }
+    public void setY(double y) { Y = y; }
 }
